Move player screen lock handling to extra class
