+++ app/src/main/java/org/jellyfin/mobile/player/PlaybackMenus.kt
@@ -47,7 +47,7 @@ class PlaybackMenus(
             fragment.onSkipToNext()
         }
         lockScreenButton.setOnClickListener {
-            fragment.lockScreen()
+            fragment.playerLockScreenHelper.lockScreen()
         }
         audioStreamsButton.setOnClickListener {
             fragment.suppressControllerAutoHide(true)
